DL- Male ispravke u bazi podataka

+++ PresentationLayer/DataLayer/Models/Picture.cs
@@ -13,7 +13,7 @@ namespace DataLayer.Models
         public string Creator { get; set; }
         public string Description { get; set; }
         public byte[] Image { get; set; }
-        public double AverageGrade { get; set; }
+        public decimal AverageGrade { get; set; }
         public int GalleryId { get; set; }
     }
 }
