+++ PresentationLayer/DataLayer/Models/Review.cs
@@ -10,7 +10,7 @@ namespace DataLayer.Models
     {
         public int Id { get; set; }
         public string Comment { get; set; }
-        public double Grade { get; set; }
+        public decimal Grade { get; set; }
         public int ReviewerId { get; set; }
         public int PictureId { get; set; }
     }
